airbyte-integrations/connectors/destination-langchain-ingest/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "destination-langchain-ingest"
description = "Destination implementation for langchain-ingest."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/destinations/langchain-ingest"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "destination_langchain_ingest" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
airbyte-cdk = "^6.33.0"
langchain_core = "0.1.42"

[tool.poetry.scripts]
destination-langchain-ingest = "destination_langchain_ingest.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"

